jailbreak_prompt_uncovering/requirements.txt
python-Levenshtein==0.12.0
sentence-transformers==0.4.1.2
scikit-learn==0.23.2
